some more mapped classes, added missing enum conversions, use counted pointers instead of ususal, first iteration of dispose implementation, use task pattern for custom scheme handlers, fixed warnings

+++ CefSharp/CefBrowser.cs
@@ -7,13 +7,13 @@ namespace CefSharp
 {
     public abstract class CefBrowserBase : ObjectBase
     {
-        public int BrowserId { get; set; }
+        public long BrowserId { get; set; }
 
         protected CefBrowserBase()
         {
         }
 
-        public abstract object EvaluateScript(int frameId, string script, double timeout);
+        public abstract object EvaluateScript(long frameId, string script, double timeout);
 
     }
 }
